% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{copyOnwriteProp}
\alias{copyOnwriteProp}
\alias{setCopyOnwrite}
\alias{unsetCopyOnwrite}
\title{Get or set the copy-on-write feature}
\usage{
copyOnwriteProp(x)

setCopyOnwrite(x)

unsetCopyOnwrite(x)
}
\arguments{
\item{x}{A shared object}
}
\value{
'copyOnwriteProp': A copy-on-write property

'setCopyOnwrite': No return value

'unsetCopyOnwrite': No return value
}
\description{
Get or set the copy-on-write feature for a shared object
}
