% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcesManager.R
\name{getDataInfo}
\alias{getDataInfo}
\title{Get a summary report of the data in the shared memory}
\usage{
getDataInfo(data_ids = NULL)
}
\arguments{
\item{data_ids}{The data ID. If NULL(default), all data will be returned.}
}
\value{
A data.frame
}
\description{
This function will return a list of the summary of the data in the shared memory.
}
\examples{
getDataInfo()
}
