% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedObject-method.R
\name{getSharedProperties}
\alias{getSharedProperties}
\title{get a summary report for a shared object}
\usage{
getSharedProperties(x, as.list = FALSE)
}
\arguments{
\item{x}{A shared object}

\item{as.list}{Logical, Whether the result should be a list of shared properties even
when only one sharedProperty object is reported.}
}
\value{
If x is not a shared object, return \code{NULL}.

If x is a shared vector and as.list is \code{FALSE}, the function returns a sharedProperty object.

Otherwise, the function returns a list of sharedProperty object(s).
}
\description{
The function reports the properties related to the shared object. the individual property
can be accessed via \code{$} operator.
}
\examples{
x = share(1:10)
getSharedProperties(x)
}
\seealso{
getCopyOnWrite, getSharedSubset, getSharedCopy
}
