% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listSharedObject}
\alias{listSharedObject}
\title{Get the shared object usage report}
\usage{
listSharedObject()
}
\value{
A data.frame object with shared object id and size
}
\description{
Get the shared object usage report
}
