% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgconfig}
\alias{pkgconfig}
\title{Find path of the shared memory header file}
\usage{
pkgconfig(x)
}
\arguments{
\item{x}{Character, "PKG_LIBS" or "PKG_CPPFLAGS"}
}
\value{
path to the header or compiler flags
}
\description{
This function will return the path of the shared memory header or
the flags that are used to compile the package
for the developers who want to use C++ level implementation of the
\code{SharedObject} package
}
\examples{
SharedObject:::pkgconfig("PKG_LIBS")
SharedObject:::pkgconfig("PKG_CPPFLAGS")
}
