% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcesManager.R
\name{removeAllObject}
\alias{removeAllObject}
\title{Release all objects in the shared memory}
\usage{
removeAllObject()
}
\value{
no return value
}
\description{
This function will force the package to delete all data in the shared memory.
Any try to read the data after the function call will crash R.
}
