% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcesManager.R
\name{removeObject}
\alias{removeObject}
\title{Release an objects in the shared memory}
\usage{
removeObject(data_ids)
}
\arguments{
\item{data_ids}{The data ID you want to delete}
}
\value{
no return value
}
\description{
This function will delete the data associated with the key provided by the function argument.
Any try to read the data after the function call will crash R.
}
