% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedObject-altrep.R
\name{sharedObject}
\alias{sharedObject}
\title{Create an R object in the shared memory}
\usage{
sharedObject(x, copyOnWrite = sharedParms.copyOnWrite(),
  sharedSub = sharedParms.sharedSub())
}
\arguments{
\item{x}{An R object that you want to shared. The data type can be `vector`,
`matrix` and `data.frame`. List is not supported.}

\item{copyOnWrite}{Whether a new R object need to be allocated when the shared object is changed.
The default value is `TRUE`, but can be changed through `sharedParms.copyOnWrite()`}

\item{sharedSub}{whether the subset of a shared object is still a shared object.
The default value is `TRUE`, but can be changed through `sharedParms.sharedSub()`}
}
\value{
A shared object
}
\description{
This function will create an object in the shared memory for the function argument [x],
It can be exported to the other R processes. If exported, the object in all processes will use
the same data in the same memory location. Therefore, there is no cost to share the return value of this function.
}
