% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgSetting.R
\name{setSharedObjectOptions}
\alias{setSharedObjectOptions}
\alias{getSharedObjectOptions}
\title{Get or set the global options for the SharedObject package}
\usage{
setSharedObjectOptions(...)

getSharedObjectOptions(...)
}
\arguments{
\item{...}{`setSharedObjectOptions`: the options you want to set, it can be copyOnWrite,
sharedSubset and sharedCopy.

`getSharedObjectOptions`: it is a character vector. If empty, all options will be returned.}
}
\description{
Get or set the global options for the SharedObject package
}
