% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedObject-altrep.R
\name{getDataID}
\alias{getDataID}
\alias{getTypeName}
\alias{getOwnData}
\alias{getProcessID}
\alias{getTypeID}
\alias{getLength}
\alias{getTotalSize}
\alias{getCopyOnWrite}
\alias{getSharedSubset}
\alias{getSharedCopy}
\alias{setCopyOnWrite}
\alias{setSharedSubset}
\alias{setSharedCopy}
\title{Get or set the properties of a shared object}
\usage{
getDataID(x)

getTypeName(x)

getOwnData(x)

getProcessID(x)

getTypeID(x)

getLength(x)

getTotalSize(x)

getCopyOnWrite(x)

getSharedSubset(x)

getSharedCopy(x)

setCopyOnWrite(x, value)

setSharedSubset(x, value)

setSharedCopy(x, value)
}
\arguments{
\item{x}{A shared object}

\item{value}{logical, the value you want the property to be. For a `data.frame`, it can be either
a single value or a vector. If the length of `value` does not match the column of the `data.frame`,
`value` will be replicate to match the length.}
}
\value{
`get`: The property of a shared object

`set`: No return value
}
\description{
Get or set the properties of a shared object
}
