% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedObject-altrep.R
\name{sharedVector}
\alias{sharedVector}
\title{A fundamental tool to share an R object.}
\usage{
sharedVector(x, opt)
}
\arguments{
\item{x}{An automic object}

\item{opt}{options}
}
\description{
A fundamental tool to share an R object.
}
