% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedObject-altrep.R
\name{sharedVectorById}
\alias{sharedVectorById}
\title{A fundamental tool to share an R object.}
\usage{
sharedVectorById(did)
}
\arguments{
\item{did}{The data id}
}
\description{
retrieve an R object by the data ID, this function is for internal usage only
}
