% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedObject-altrep.R
\name{getVecDID}
\alias{getVecDID}
\alias{getVecPID}
\alias{getVecType}
\alias{getVecTypeID}
\alias{getVecTotalSize}
\alias{getVecCopyOnWrite}
\alias{getVecSharedSub}
\alias{getVecSharedDuplicate}
\alias{setVecCopyOnwrite}
\alias{setVecSharedSub}
\alias{setVecSharedDuplicate}
\alias{setVecOwnData}
\title{Get/Set the property of a shared vector}
\usage{
getVecDID(x)

getVecPID(x)

getVecType(x)

getVecTypeID(x)

getVecTotalSize(x)

getVecCopyOnWrite(x)

getVecSharedSub(x)

getVecSharedDuplicate(x)

setVecCopyOnwrite(x, value)

setVecSharedSub(x, value)

setVecSharedDuplicate(x, value)

setVecOwnData(x, value)
}
\arguments{
\item{x}{a shared vector object}

\item{value}{the value you want to change}
}
\description{
These function works for atomic shared vector only, it does not work for data.frame and list type.
}
