% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share.R
\name{tryShare}
\alias{tryShare}
\title{#' @rdname share
#' @export
setMethod("share", signature(x = "character"), promptError)
#' @rdname share
#' @export
setMethod("share", signature(x = "vector"), shareAtomic)
#' @rdname share
#' @export
setMethod("share", signature(x = "matrix"), shareAtomic)
#' @rdname share
#' @export
setMethod("share", signature(x = "list"), shareList)
setMethod("share", signature(x = "environment"), promptError)}
\usage{
tryShare(x, ...)
}
\description{
#' @rdname share
#' @export
setMethod("share", signature(x = "character"), promptError)
#' @rdname share
#' @export
setMethod("share", signature(x = "vector"), shareAtomic)
#' @rdname share
#' @export
setMethod("share", signature(x = "matrix"), shareAtomic)
#' @rdname share
#' @export
setMethod("share", signature(x = "list"), shareList)
setMethod("share", signature(x = "environment"), promptError)
}
