% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedObject-constructor.R
\name{unshare}
\alias{unshare}
\alias{unshare,ANY-method}
\alias{unshare,vector-method}
\alias{unshare,list-method}
\title{Unshare a shared object}
\usage{
unshare(x)
}
\arguments{
\item{x}{a shared object, or an object that contains a shared object.}
}
\value{
An unshared object
}
\description{
Unshare a shared object. There will be no effect if the
object is not shared.
}
\examples{
x1 <- share(1:10)
x2 <- unshare(x1)
is.shared(x1)
is.shared(x2)
}
