% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{setCopyOnwrite}
\alias{setCopyOnwrite}
\alias{unsetCopyOnwrite}
\title{Set the copy-on-write feature}
\usage{
setCopyOnwrite(x)

unsetCopyOnwrite(x)
}
\arguments{
\item{x}{A shared object}
}
\value{
'setCopyOnwrite': No return value

'unsetCopyOnwrite': No return value
}
\description{
Set the copy-on-write feature for a shared object
}
