% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getSharedProperty}
\alias{getSharedProperty}
\title{get a summary report for a shared object}
\usage{
getSharedProperty(x, as.list = FALSE)
}
\arguments{
\item{x}{A shared object}

\item{as.list}{Logical, Whether the result should be in a list format even
if only one shared property is reported}
}
\description{
get a summary report for a shared object
}
