% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getSharedProperty}
\alias{getSharedProperty}
\title{get a summary report for a shared object}
\usage{
getSharedProperty(x, as.list = FALSE)
}
\arguments{
\item{x}{A shared object}
}
\description{
get a summary report for a shared object
}
