% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcesManager.R, R/sharedObject-altrep.R,
%   R/tools.R
\name{.removeAllObject}
\alias{.removeAllObject}
\alias{.removeObject}
\alias{.sharedVectorById}
\alias{.dataID}
\alias{.typeName}
\alias{.ownData}
\alias{.processID}
\alias{.typeID}
\alias{.length}
\alias{.totalSize}
\alias{.createInheritedParms}
\title{Internal functions}
\usage{
.removeAllObject()

.removeObject(dataID)

.sharedVectorById(did)

.dataID(x)

.typeName(x)

.ownData(x)

.processID(x)

.typeID(x)

.length(x)

.totalSize(x)

.createInheritedParms(x)
}
\arguments{
\item{dataID}{The data ID you want to delete}

\item{did}{The data id}

\item{x}{A shared object}
}
\value{
.removeAllObject: no return value

.removeObject: no return value
}
\description{
Internal functions
}
\details{
.removeAllObject: This function will force the package to delete all data in the shared memory.
Any try to read the data after the function call will crash R.

removeObject: This function will delete the data associated with the key provided by the function argument.
Any try to read the data after the function call will crash R.

.sharedVectorById: retrieve an R object by the data ID

.createInheritedParms: Get the parameters that can be inherited by the child of a shared object
}
