% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcesManager.R, R/sharedObject-method.R
\name{removeAllObject}
\alias{removeAllObject}
\alias{removeObject}
\alias{.getProperty}
\alias{.setProperty}
\title{Internal functions}
\usage{
removeAllObject()

removeObject(dataId)

.getProperty(x, property)

.setProperty(x, property, value)
}
\arguments{
\item{dataId}{The data ID you want to delete}

\item{x}{A shared object}

\item{property}{Character, the name of the property}

\item{value}{The new value of the property}
}
\value{
removeAllObject: no return value

removeObject: no return value
}
\description{
Internal functions
}
\details{
removeAllObject: This function will force the package to delete all data in the shared memory.
Any try to read the data after the function call will crash R.

removeObject: This function will delete the data associated with the key provided by the function argument.
Any try to read the data after the function call will crash R.
}
\examples{
removeAllObject()
}
