% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.altrep}
\alias{is.altrep}
\title{Whether an object is an ALTREP object}
\usage{
is.altrep(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Whether an object is an ALTREP object
}
\examples{
x <- share(runif(10))
is.altrep(x)
}
