% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedObject-method.R
\name{is.shared}
\alias{is.shared}
\alias{is.shared,ANY-method}
\alias{is.shared,list-method}
\title{Test whether the object is a shared object}
\usage{
is.shared(x, ...)

\S4method{is.shared}{ANY}(x, ...)

\S4method{is.shared}{list}(x, ...)
}
\arguments{
\item{x}{An R object}

\item{...}{For generalization purpose only}
}
\value{
TRUE/FALSE indicating whether the object is a shared object.
If the object is a list, the return value is a vector of TRUE/FALSE corresponding
to each element of the list.
}
\description{
Test whether the object is a shared object
}
\examples{
x <- share(1:10)
is.shared(x)
}
