% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedObject-method.R
\name{is.shared}
\alias{is.shared}
\alias{is.shared.default}
\alias{is.shared.list}
\alias{is.shared.data.frame}
\title{Test whether the object is a shared object}
\usage{
is.shared(x)

\method{is.shared}{default}(x)

\method{is.shared}{list}(x)

\method{is.shared}{data.frame}(x)
}
\arguments{
\item{x}{An R object}
}
\value{
TRUE/FALSE indicating whether the object is a shared object.
If the object is a list, the return value is a vector of TRUE/FALSE corresponding
to each element of the list.
}
\description{
Test whether the object is a shared object
}
