% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developerAPIS.R
\name{listSharedObject}
\alias{listSharedObject}
\title{Get the shared object usage report}
\usage{
listSharedObject(end = NULL, start = NULL, includeCharId = FALSE)
}
\arguments{
\item{end}{the end value of the ID. The default is \code{NULL}. See details.}

\item{start}{the start value of the ID. The default is \code{NULL}. See details.}

\item{includeCharId}{Whether including the shared objects named by a character ID, it only works
on Unix-like systems. See \code{?allocateNamedSharedMemory} for more information. The default is \code{FALSE}.}
}
\value{
A data.frame object with shared object id and size
}
\description{
Get the shared object usage report. The size is the real memory size
that a system allocates for the shared object, so it might be larger
than the object size. The size unit is byte.
}
\details{
The parameter \code{start} and \code{end} specify the range of the ID. If not specified, all
IDs will be listed.

On Ubuntu or many other Unix-like operating systems, the shared objects
can be found in the folder \verb{/dev/shm}. The function can find all shared objects
if the folder exists.

On Windows, since there is no easy way to find all shared objects.
the function will guess the range of the shared object IDs and search all IDs
within the range. Therefore, if there are too many shared objects(over 4 billions)
,the object id can be out of the searching range and the result may not be complete.
Furthermore, there will be no named shared object in the returned list.
}
\examples{
## Automatically determine the search range
listSharedObject()

## specify the search range
listSharedObject(start = 10, end = 20)

## Search from 0 to 20
listSharedObject(20)
}
\seealso{
\code{\link{getLastIndex}}, \code{\link{allocateSharedMemory}},
\code{\link{allocateNamedSharedMemory}}, \code{\link{mapSharedMemory}}, \code{\link{unmapSharedMemory}},
\code{\link{freeSharedMemory}}, \code{\link{hasSharedMemory}}, \code{\link{getSharedMemorySize}}
}
