% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-options.R
\name{setSharedObjectOptions}
\alias{setSharedObjectOptions}
\alias{getSharedObjectOptions}
\title{Get or set the global options for the SharedObject package}
\usage{
setSharedObjectOptions(...)

getSharedObjectOptions(...)
}
\arguments{
\item{...}{\code{setSharedObjectOptions}: the options you want to set, it can be copyOnWrite,
sharedSubset and sharedCopy.

\code{getSharedObjectOptions}: A character vector. If empty, all options will be returned.}
}
\value{
\code{setSharedObjectOptions}: No return value
\code{getSharedObjectOptions}: A list of the package options or a single value
}
\description{
Get or set the global options for the SharedObject package
}
\examples{
getSharedObjectOptions()
setSharedObjectOptions(copyOnWrite = FALSE)
getSharedObjectOptions()
getSharedObjectOptions("copyOnWrite")


}
