% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/sharedObjectProperty.R
\name{getSharedObjectProperty}
\alias{getSharedObjectProperty}
\alias{setSharedObjectProperty}
\alias{getSharedObjectProperty,ANY,characterOrNULLOrMissing-method}
\alias{getSharedObjectProperty,list,characterOrNULLOrMissing-method}
\alias{setSharedObjectProperty,ANY,characterOrNULLOrMissing-method}
\alias{setSharedObjectProperty,list,characterOrNULLOrMissing-method}
\alias{getCopyOnWrite}
\alias{getSharedSubset}
\alias{getSharedCopy}
\alias{setCopyOnWrite}
\alias{setSharedSubset}
\alias{setSharedCopy}
\title{Get/Set the properties of the shared object.}
\usage{
getSharedObjectProperty(x, property, ...)

setSharedObjectProperty(x, property, value, ...)

\S4method{getSharedObjectProperty}{ANY,characterOrNULLOrMissing}(x, property, ...)

\S4method{getSharedObjectProperty}{list,characterOrNULLOrMissing}(x, property, ...)

\S4method{setSharedObjectProperty}{ANY,characterOrNULLOrMissing}(x, property, value, ...)

\S4method{setSharedObjectProperty}{list,characterOrNULLOrMissing}(x, property, value, ...)

getCopyOnWrite(x)

getSharedSubset(x)

getSharedCopy(x)

setCopyOnWrite(x, value)

setSharedSubset(x, value)

setSharedCopy(x, value)
}
\arguments{
\item{x}{A shared object}

\item{property}{A character vector, the name of the property(s),
if the argument is missing or the value is \code{NULL}, it represents all properties.}

\item{...}{Not used}

\item{value}{The new value of the property, if the length of value
does not match the length of the property, the argument \code{value} will
be repeated to match the length.}
}
\value{
get: The property(s) of a shared object

set: No return value
}
\description{
Get/Set the properties of the shared object.
The available properties are \code{dataId}, \code{length}, \code{totalSize},
\code{dataType}, \code{ownData}, \code{copyOnWrite}, \code{sharedSubset}, \code{sharedCopy}.
}
\examples{
x = share(1:20)

## Check the default values
getSharedObjectProperty(x, NULL)
getCopyOnWrite(x)
getSharedSubset(x)
getSharedCopy(x)

## Set the values
setCopyOnWrite(x, FALSE)
setSharedSubset(x, FALSE)
setSharedCopy(x, TRUE)

## Check the values again
getSharedObjectProperty(x, NULL)
getCopyOnWrite(x)
getSharedSubset(x)
getSharedCopy(x)
}
