% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedObject-method.R
\name{getCopyOnWrite}
\alias{getCopyOnWrite}
\alias{getSharedSubset}
\alias{getSharedCopy}
\alias{setCopyOnWrite}
\alias{setSharedSubset}
\alias{setSharedCopy}
\title{Get or set the properties of a shared object}
\usage{
getCopyOnWrite(x)

getSharedSubset(x)

getSharedCopy(x)

setCopyOnWrite(x, value)

setSharedSubset(x, value)

setSharedCopy(x, value)
}
\arguments{
\item{x}{A shared object}

\item{value}{logical, the value you want the property to be. For a \code{data.frame}, it can be either
a single value or a vector. If the length of \code{value} does not match the column of the \code{data.frame},
\code{value} will be replicate to match the length.}
}
\value{
\code{get}: The property of a shared object

\code{set}: No return value
}
\description{
Get or set the properties of a shared object
}
\examples{
x=share(1:20)
##Check the default values
getCopyOnWrite(x)
getSharedSubset(x)
getSharedCopy(x)

##Set the values
setCopyOnWrite(x,FALSE)
setSharedSubset(x,FALSE)
setSharedCopy(x,TRUE)

##Check the values again
getCopyOnWrite(x)
getSharedSubset(x)
getSharedCopy(x)
}
