% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedObject-altrep.R
\name{sharedVector}
\alias{sharedVector}
\title{A fundamental tool to share an R object.}
\usage{
sharedVector(x, opt = list())
}
\arguments{
\item{x}{An automic object}

\item{opt}{options}
}
\description{
This function only handles an automic object and does not attach the object's attributes, Please call the function `sharedObject` instead
}
