% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is.altrep}
\alias{is.altrep}
\alias{is.shared}
\title{Is an Object a desired type?}
\usage{
is.altrep(x)

is.shared(x, recursive = TRUE)
}
\arguments{
\item{x}{an R object}

\item{recursive}{Logical, whether a `data.frame` can be treated as a shared object. If `TRUE`, a `data.frame`
is called a shared object if and only if all of its columns are shared objects.}
}
\description{
Is an Object a desired type?
}
